+++ src/Widget/TwigAware.php
@@ -7,8 +7,8 @@ namespace Bolt\Widget;
 use Twig\Environment;
 
 /**
- * Interface TwigAware - Widgets that make use of a template to render their
- * contents need to implement this interface, in order to access the current
+ * Trait TwigAware - Widgets that make use of a template to render their
+ * contents need to use this Trait, in order to access the current
  * Twig\Environment and to render a Twig template.
  */
 trait TwigAware
