+++ src/Widget/WidgetInterface.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Bolt\Widget;
 
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Twig\Environment;
 
 interface WidgetInterface
 {
